first changes, work in progress
