+++ sktime/benchmarking/__init__.py
@@ -1,0 +1,7 @@
+from sktime.benchmarking._io import DatasetHDD, Result, ResultRAM, ResultHDD
+from sktime.benchmarking._io import DatasetRAM
+from sktime.benchmarking._io import DataHDD
+from sktime.benchmarking.metrics import ScoreAccuracy
+from sktime.benchmarking.metrics import ScoreMSE
+from sktime.benchmarking.orchestration import Orchestrator
+from sktime.benchmarking.evaluation import Evaluator
